+++ lib/karafka/instrumentation/vendors/datadog/listener.rb
@@ -46,7 +46,11 @@ module Karafka
             RdKafkaMetric.new(:count, :brokers, 'connection.disconnects', 'disconnects_d'),
             RdKafkaMetric.new(:gauge, :brokers, 'network.latency.avg', %w[rtt avg]),
             RdKafkaMetric.new(:gauge, :brokers, 'network.latency.p95', %w[rtt p95]),
-            RdKafkaMetric.new(:gauge, :brokers, 'network.latency.p99', %w[rtt p99])
+            RdKafkaMetric.new(:gauge, :brokers, 'network.latency.p99', %w[rtt p99]),
+
+            # Topics metrics
+            RdKafkaMetric.new(:gauge, :topics, 'consumer.lags', 'consumer_lag_stored'),
+            RdKafkaMetric.new(:gauge, :topics, 'consumer.lags_delta', 'consumer_lag_stored_d')
           ].freeze
 
           configure
@@ -68,9 +72,12 @@ module Karafka
           # @param event [Karafka::Core::Monitoring::Event]
           def on_statistics_emitted(event)
             statistics = event[:statistics]
+            consumer_group_id = event[:consumer_group_id]
+
+            base_tags = default_tags + ["consumer_group:#{consumer_group_id}"]
 
             rd_kafka_metrics.each do |metric|
-              report_metric(metric, statistics)
+              report_metric(metric, statistics, base_tags)
             end
           end
 
@@ -83,10 +90,7 @@ module Karafka
             if event.payload[:caller].respond_to?(:messages)
               metadata = event.payload[:caller].messages.metadata
 
-              extra_tags += [
-                "topic:#{metadata.topic}",
-                "partition:#{metadata.partition}"
-              ]
+              extra_tags += consumer_tags(event.payload[:caller])
             end
 
             count('error_occurred', 1, tags: default_tags + extra_tags)
@@ -99,21 +103,23 @@ module Karafka
             time_taken = event[:time]
             messages_count = event[:messages_buffer].size
 
-            histogram('listener.polling.time_taken', time_taken, tags: default_tags)
-            histogram('listener.polling.messages', messages_count, tags: default_tags)
+            consumer_group_id = event[:subscription_group].consumer_group_id
+
+            extra_tags = ["consumer_group:#{consumer_group_id}"]
+
+            histogram('listener.polling.time_taken', time_taken, tags: default_tags + extra_tags)
+            histogram('listener.polling.messages', messages_count, tags: default_tags + extra_tags)
           end
 
           # Here we report majority of things related to processing as we have access to the
           # consumer
           # @param event [Karafka::Core::Monitoring::Event]
           def on_consumer_consumed(event)
-            messages = event.payload[:caller].messages
+            consumer = event.payload[:caller]
+            messages = consumer.messages
             metadata = messages.metadata
 
-            tags = default_tags + [
-              "topic:#{metadata.topic}",
-              "partition:#{metadata.partition}"
-            ]
+            tags = default_tags + consumer_tags(consumer)
 
             count('consumer.messages', messages.count, tags: tags)
             count('consumer.batches', 1, tags: tags)
@@ -126,26 +132,14 @@ module Karafka
 
           # @param event [Karafka::Core::Monitoring::Event]
           def on_consumer_revoked(event)
-            messages = event.payload[:caller].messages
-            metadata = messages.metadata
-
-            tags = default_tags + [
-              "topic:#{metadata.topic}",
-              "partition:#{metadata.partition}"
-            ]
+            tags = default_tags + consumer_tags(event.payload[:caller])
 
             count('consumer.revoked', 1, tags: tags)
           end
 
           # @param event [Karafka::Core::Monitoring::Event]
           def on_consumer_shutdown(event)
-            messages = event.payload[:caller].messages
-            metadata = messages.metadata
-
-            tags = default_tags + [
-              "topic:#{metadata.topic}",
-              "partition:#{metadata.partition}"
-            ]
+            tags = default_tags + consumer_tags(event.payload[:caller])
 
             count('consumer.shutdown', 1, tags: tags)
           end
@@ -198,14 +192,15 @@ module Karafka
           # Reports a given metric statistics to Datadog
           # @param metric [RdKafkaMetric] metric value object
           # @param statistics [Hash] hash with all the statistics emitted
-          def report_metric(metric, statistics)
+          # @param base_tags [Array<String>] base tags we want to start with
+          def report_metric(metric, statistics, base_tags)
             case metric.scope
             when :root
               public_send(
                 metric.type,
                 metric.name,
                 statistics.fetch(*metric.key_location),
-                tags: default_tags
+                tags: base_tags
               )
             when :brokers
               statistics.fetch('brokers').each_value do |broker_statistics|
@@ -218,12 +213,46 @@ module Karafka
                   metric.type,
                   metric.name,
                   broker_statistics.dig(*metric.key_location),
-                  tags: default_tags + ["broker:#{broker_statistics['nodename']}"]
+                  tags: base_tags + ["broker:#{broker_statistics['nodename']}"]
                 )
+              end
+            when :topics
+              statistics.fetch('topics').each do |topic_name, topic_values|
+                topic_values['partitions'].each do |partition_name, partition_statistics|
+                  next if partition_name == '-1'
+                  # Skip until lag info is available
+                  next if partition_statistics['consumer_lag'] == -1
+
+                  public_send(
+                    metric.type,
+                    metric.name,
+                    partition_statistics.dig(*metric.key_location),
+                    tags: base_tags + [
+                      "topic:#{topic_name}",
+                      "partition:#{partition_name}"
+                    ]
+                  )
+                end
               end
             else
               raise ArgumentError, metric.scope
             end
+          end
+
+          # Builds basic per consumer tags for publication
+          #
+          # @param consumer [Karafka::BaseConsumer]
+          # @return [Array<String>]
+          def consumer_tags(consumer)
+            messages = consumer.messages
+            metadata = messages.metadata
+            consumer_group_id = consumer.topic.consumer_group.id
+
+            [
+              "topic:#{metadata.topic}",
+              "partition:#{metadata.partition}",
+              "consumer_group:#{consumer_group_id}"
+            ]
           end
         end
       end